+++ server/controllers/LoaiSan.controller.js
@@ -1,5 +1,6 @@
 const LoaiSanService = require("../services/LoaiSan.service");
 const successHandler = require("../utils/successHandler");
+const errorHandler = require("../utils/errorHandler");
 
 const getAllLoaiSan = async (req, res, next) => {
   try {
@@ -16,6 +17,7 @@ const getLoaiSanById = async (req, res, next) => {
 
   try {
     const result = await LoaiSanService.getLoaiSanById(MaLoaiSan);
+    if (!result) return errorHandler(res, "KHÔNG TÌM THẤY LOẠI SÂN", 403);
 
     successHandler(res, "THÔNG TIN LOẠI SÂN!", result);
   } catch (error) {
@@ -31,7 +33,6 @@ const addLoaiSan = async (req, res, next) => {
       MoTa,
       TrangThai,
     });
-
     successHandler(res, "THÊM MỚI LOẠI SÂN THÀNH CÔNG!", result, 201);
   } catch (error) {
     next(error);
@@ -43,12 +44,14 @@ const updateLoaiSan = async (req, res, next) => {
   const { TenLoaiSan, MoTa, TrangThai } = req.body;
 
   try {
+    const findId = await LoaiSanService.findLoaiSanById(MaLoaiSan);
+    if (!findId) return errorHandler(res, "KHÔNG TÌM THẤY LOẠI SÂN", 403);
+
     const result = await LoaiSanService.updateLoaiSan(MaLoaiSan, {
       TenLoaiSan,
       MoTa,
       TrangThai,
     });
-
     successHandler(res, "CẬP NHẬT LOẠI SÂN THÀNH CÔNG!", result);
   } catch (error) {
     next(error);
@@ -59,8 +62,10 @@ const deleteLoaiSan = async (req, res, next) => {
   const { MaLoaiSan } = req.params;
 
   try {
-    const result = await LoaiSanService.deleteLoaiSanXoa(MaLoaiSan);
+    const findId = await LoaiSanService.findLoaiSanById(MaLoaiSan);
+    if (!findId) return errorHandler(res, "KHÔNG TÌM THẤY LOẠI SÂN", 403);
 
+    const result = await LoaiSanService.deleteLoaiSanXoa(MaLoaiSan);
     successHandler(res, "XÓA LOẠI SÂN THÀNH CÔNG!", result);
   } catch (error) {
     next(error);
